New Data Types & Removal of now useless code
Introductions:
🟢 Introduction of new Data Types.

Changes:
🔵 Old Data Types are now compatible with the new Sensor Base Class.
🔵 Removed of bugged code from now useless expansions.
🔵 Fixed Orion Redefinition of constants.

Known issues:
🔴 Panic function must be preloaded before Panic is called because the system might panic because there is not enough ram.
🔴 The system might panic from time to time.
🔴 FileBlob does not store changes.

Future improvements:
🟡 Add more usefull tools for the ConsoleLogger.
🟡 Add specific GPS Module.
🟡 Add support for specific Camera Module.
🟡 Optimize libs.
🟡 Introduce Teensy Thread Manager and General Thread Manager
🟡 Will introduce TSL2591 module.

+++ Orion/Sensors/Sensor.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Orion/Orion.h"
+#include "Orion/Orion.hpp"
 
 namespace Orion {
     namespace Sensors {
@@ -9,7 +9,7 @@ namespace Orion {
                 bool _initialized = false;
                 uint8_t _state = 0x00;
                 uint64_t _type = 0x00;
-                uint32_t _lastUpdateTime = 0;
+                unsigned long _lastUpdateTime = 0;
 
                 virtual uint64_t* GetDataTypes();
                 virtual uint64_t GetDataTypesSize();
